+++ resource/interface.go
@@ -8,5 +8,5 @@ import (
 type Interface interface {
 	SaveValueReceived(value float32, date time.Time, description string, bank string)
 	SaveBankAccount(acronym, description string) error
-	SelectBanksAccountsByUserID(id int64) ([]*bankaccounts.BankAccount, error)
+	SelectBanksAccountsByUserID(id int) ([]*bankaccounts.BankAccount, error)
 }
